rename admin.ts to user.ts
